internal/networkrules: move ParseRule into a separate file
