+++ internal/networkrules/networkrules.go
@@ -4,17 +4,10 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
-	"regexp"
-	"strings"
 
 	"github.com/irbis-sh/zen-desktop/internal/networkrules/exceptionrule"
 	"github.com/irbis-sh/zen-desktop/internal/networkrules/rule"
 	"github.com/irbis-sh/zen-desktop/internal/ruletree"
-)
-
-var (
-	reHosts       = regexp.MustCompile(`^(?:0\.0\.0\.0|127\.0\.0\.1)\s(.+)`)
-	reHostsIgnore = regexp.MustCompile(`^(?:0\.0\.0\.0|broadcasthost|local|localhost(?:\.localdomain)?|ip6-\w+)$`)
 )
 
 type ruleStore[T any] interface {
@@ -33,80 +26,6 @@ func New() *NetworkRules {
 		primaryStore:   ruletree.New[*rule.Rule](),
 		exceptionStore: ruletree.New[*exceptionrule.ExceptionRule](),
 	}
-}
-
-func (nr *NetworkRules) ParseRule(rawRule string, filterName *string) (isException bool, err error) {
-	if matches := reHosts.FindStringSubmatch(rawRule); matches != nil {
-		hostsField := matches[1]
-		if commentIndex := strings.IndexByte(hostsField, '#'); commentIndex != -1 {
-			hostsField = hostsField[:commentIndex]
-		}
-
-		// An IP address may be followed by multiple hostnames.
-		//
-		// As stated in https://man.freebsd.org/cgi/man.cgi?hosts(5):
-		// "Items are separated by any number of blanks and/or tab characters."
-		hosts := strings.Fields(hostsField)
-
-		for _, host := range hosts {
-			if reHostsIgnore.MatchString(host) {
-				continue
-			}
-
-			pattern := fmt.Sprintf("||%s^", host)
-			nr.primaryStore.Insert(pattern, &rule.Rule{
-				RawRule:    rawRule,
-				FilterName: filterName,
-				Document:   true,
-			})
-		}
-
-		return false, nil
-	}
-
-	if strings.HasPrefix(rawRule, "@@") {
-		r := &exceptionrule.ExceptionRule{
-			RawRule:    rawRule,
-			FilterName: filterName,
-		}
-
-		pattern, modifiers, found := strings.Cut(rawRule[2:], "$")
-		if pattern != "" && pattern[0] == '/' && pattern[len(pattern)-1] == '/' {
-			// This is a regexp rule.
-			// TODO: implement proper support for regexp rules.
-			return true, nil
-		}
-		if found {
-			modifiersArr := splitModifiers(modifiers)
-			if err := r.ParseModifiers(modifiersArr); err != nil {
-				return false, fmt.Errorf("parse modifiers: %v", err)
-			}
-		}
-		nr.exceptionStore.Insert(pattern, r)
-
-		return true, nil
-	}
-
-	r := &rule.Rule{
-		RawRule:    rawRule,
-		FilterName: filterName,
-	}
-
-	pattern, modifiers, found := strings.Cut(rawRule, "$")
-	if pattern != "" && pattern[0] == '/' && pattern[len(pattern)-1] == '/' {
-		// This is a regexp rule.
-		// TODO: implement proper support for regexp rules.
-		return false, nil
-	}
-	if found {
-		modifiersArr := splitModifiers(modifiers)
-		if err := r.ParseModifiers(modifiersArr); err != nil {
-			return false, fmt.Errorf("parse modifiers: %v", err)
-		}
-	}
-	nr.primaryStore.Insert(pattern, r)
-
-	return false, nil
 }
 
 func (nr *NetworkRules) ModifyReq(req *http.Request) (appliedRules []rule.Rule, shouldBlock bool, redirectURL string) {
@@ -234,42 +153,4 @@ func renderURLWithoutPort(u *url.URL) string {
 	}
 
 	return stripped.String()
-}
-
-// splitModifiers splits by unescaped commas.
-// Empty fields are preserved (like strings.Split).
-func splitModifiers(s string) []string {
-	var res []string
-	var b strings.Builder
-	escaped := false
-
-	for _, r := range s {
-		switch r {
-		case '\\':
-			if escaped {
-				b.WriteRune('\\')
-			}
-			escaped = !escaped
-		case ',':
-			if escaped {
-				b.WriteRune(',')
-				escaped = false
-			} else {
-				res = append(res, b.String())
-				b.Reset()
-			}
-		default:
-			if escaped {
-				b.WriteRune('\\')
-				escaped = false
-			}
-			b.WriteRune(r)
-		}
-	}
-
-	if escaped {
-		b.WriteRune('\\')
-	}
-	res = append(res, b.String())
-	return res
 }
